[ Work-In-Progress ] Add methods for catalog for create_legacy_delta_table/add_legacy_delta_partition
Associated issue:
https://github.com/awslabs/aws-data-wrangler/issues/1055

Aim is to add two methods for creating a table using Delta lake. Link:
https://delta.io/

I've called the methods legacy (open to naming suggestions) since these are intended for systems which don't support Delta natively. These include:
* Redshift Spectrum
* Athena

These rely on a manifest for specifying which files the client needs to process and processes them like normal Parquet files. See:
https://docs.delta.io/latest/presto-integration.html
https://docs.delta.io/latest/redshift-spectrum-integration.html

Native implementations would include Spark for example which is able to read the Delta metadata and is not supported by this code as of yet.
This is very much a work in progress (no tests etc). Will let you know when it's ready for review.